measures(phase-4): split function-complexity-health into 4 single-metric tests; multi-file _shared/ primitives split
Per tasks.md Phase 4 sub-steps 4.1-4.7 (with principled deviation on 4.1):

- Split function-complexity-health.test.ts into 4 single-metric files
  under packages/measures/src/health/complexity/:
    cyclomatic-complexity.test.ts  (MAX_CYCLOMATIC_COMPLEXITY = 50)
    maintainability-index.test.ts  (MIN_MAINTAINABILITY_INDEX = 0)
    crap0.test.ts                  (MAX_CRAP_ZERO_COVERAGE = 2500)
    runtime-fan-in.test.ts         (MAX_RUNTIME_FAN_IN = 110)
- Promote scaffolding from .test.ts-excluded scope to production-visible
  _shared/ — architectural intent of the split.
- Tasks.md 4.1 micro-spec was one file complexity-primitives.ts.
  Ivan's pre-impl diagnostic (brain 80db9d5) predicted that one file
  would have 15-25 exports, breaching exports-per-file-p90 <= 10.
  Principled deviation: split into 5 conceptually-narrow files instead
  of one large one. Fixes the budget signal at the design level rather
  than working around it.
- GATE_FILES const in gate-integrity.test.ts did not reference the
  deleted combined test at HEAD, so no source edit was needed there;
  gate-integrity still passes after the split.
- All 4 budget constants carried verbatim. All 4 report IDs preserved
  unchanged.

Verifier strategy: 3-tier explained-delta gate (Phase 3 template) against
Phase-3-snapshot baseline (brain ae5874c). Promoted functions enter
production function-* reports — predictions in
notes/phase-4-expected-deltas.json (Ivan). All named budgets predicted
GREEN; new function rows expected in cyclomatic/MI/crap/fan-in reports.
