+++ packages/measures/src/health/complexity/runtime-fan-in.test.ts
@@ -1,0 +1,34 @@
+import {describe, expect, it} from 'vitest'
+import {discoverPackages} from '../../_shared/discover-packages'
+import {discoverSourceFiles} from '../../_shared/function-discovery'
+import {recordHealthMetric} from '../../_shared/report-writer'
+import {buildRuntimeSymbolsByTarget, runtimeFanInRows} from '../../_shared/runtime-fan-in'
+
+// Captured 2026-05-15 after widening discovery to whole repo via discoverPackages(); ratchet down over time.
+const MAX_RUNTIME_FAN_IN = 110         // observed max: 107 (graph-model receives 107 named symbols)
+
+describe('runtime fan-in health', () => {
+    it('keeps runtime fan-in within budget', async () => {
+        const packages = await discoverPackages()
+        const files = await discoverSourceFiles(packages)
+        const runtimeSymbolsByTarget = await buildRuntimeSymbolsByTarget(packages, files)
+        const runtimeFanIn = runtimeFanInRows(runtimeSymbolsByTarget)
+        const maxRuntimeFanIn = runtimeFanIn[0]?.runtimeSymbols ?? 0
+
+        console.info(`\nRuntime fan-in:\n${runtimeFanIn.map(row => `${row.packageName} | ${row.runtimeSymbols} | ${row.top.join(', ')}`).join('\n')}`)
+
+        await recordHealthMetric({
+            metricId: 'runtime-fan-in',
+            metricName: 'Runtime Fan-In',
+            description: 'Maximum distinct runtime symbols imported from a package by other discovered packages.',
+            category: 'Coupling',
+            current: maxRuntimeFanIn,
+            budget: MAX_RUNTIME_FAN_IN,
+            comparison: 'lte',
+            unit: 'symbols',
+            details: {runtimeFanIn, fileCount: files.length},
+        })
+
+        expect.soft(maxRuntimeFanIn).toBeLessThanOrEqual(MAX_RUNTIME_FAN_IN)
+    }, 60000)
+})
